requirements.txt
pandas
scikit-learn
matplotlib
numpy
torch==1.9.0
kafka-python
aerospike==6.0.0
jproperties==2.1.1